update events map
